Refactor spot and traffic-log routes
