+++ app/routes/api/api.router.js
@@ -7,8 +7,8 @@ const artistRouter = require("./artist.api.router");
 const crateRouter = require("./crate/router");
 const playlistRouter = require("./playlist/router");
 const searchRouter = require("./search.api.router");
-const spotRouter = require("./spot.api.router");
-const trafficLogRouter = require("./trafficLog.router");
+const spotRouter = require("./spot/router");
+const trafficLogRouter = require("./traffic-log/router");
 const tokenRouter = require("./token.api.router");
 const { sendErrorCode } = require("./errors");
 const authorizeWithToken = passport.authenticate("jwt", { session: false });
