+++ app/routes/api/crate/router.js
@@ -1,6 +1,14 @@
 const router = require("express").Router();
 const controller = require("./controller");
-const { validateId, validateIndex, validateName, validateNewIndex, validatePath, validateItem, validateBodyIndex } = require("./validators");
+const {
+  validateId,
+  validateIndex,
+  validateName,
+  validateNewIndex,
+  validatePath,
+  validateItem,
+  validateBodyIndex,
+} = require("./validators");
 const { checkErrors } = require("../errors");
 
 router.get("/", controller.listCrates);
